payloads: Add tags to run on specific machines only
The tags will be matched against the only test machines.
This will allow reuse the existing harddisk image, increasing the
livetime of harddisk inside the test machines and also decreasing
the test duration.
In addition a smarter deployment method will be used and an
overlayfs to prevent faster degration of the used SSDs.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>